REVISED annotation job for batch processing

+++ app/jobs/text_annotation_job.rb
@@ -1,27 +1,18 @@
-class TextAnnotationJob < Struct.new(:target, :filename, :dictionaries, :options)
+class TextAnnotationJob < Struct.new(:targets, :filename, :dictionaries, :options)
   include StateManagement
 
 	def perform
     begin
+      raise ArgumentError, "The annotation targets have to be in an array." unless targets.class == Array
       annotator = TextAnnotator.new(dictionaries, options[:tokens_len_max], options[:threshold], options[:rich])
 
-      results = if target.class == Array
-        target.map do |t|
-          r = annotator.annotate(t[:text])
-          t[:denotations] = r[:denotations] if r[:denotations].present?
-          t[:relations] = r[:relations] if r[:relations].present?
-          t[:modifications] = r[:modifications] if r[:modifications].present?
-          t
-        end
-      else
-        r = annotator.annotate(target[:text])
-        target[:denotations] = r[:denotations] if r[:denotations].present?
-        target[:relations] = r[:relations] if r[:relations].present?
-        target[:modifications] = r[:modifications] if r[:modifications].present?
-        target
+      annotations_col = targets.each_slice(100).inject([]) do |col, slice|
+        col += annotator.annotate_batch(slice)
       end
 
-      File.write(TextAnnotator::RESULTS_PATH + filename + '.json', JSON.generate(results))
+      annotator.done
+
+      File.write(TextAnnotator::RESULTS_PATH + filename + '.json', JSON.generate(annotations_col))
       File.delete(TextAnnotator::RESULTS_PATH + filename)
     rescue => e
       File.write(TextAnnotator::RESULTS_PATH + filename + '.json', JSON.generate({"message":e.message}))
